test(config): use @JsonTest for URL normalization tests to get Spring's ObjectMapper
Extract URL normalization tests from SolrConfigTest into a dedicated
SolrConfigUrlNormalizationTest annotated with @JsonTest, so Spring's
auto-configured ObjectMapper is injected rather than using new ObjectMapper().
