+++ src/test/java/org/apache/solr/mcp/server/config/SolrConfigTest.java
@@ -18,13 +18,10 @@ package org.apache.solr.mcp.server.config;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.solr.client.solrj.SolrClient;
 import org.apache.solr.client.solrj.impl.Http2SolrClient;
 import org.apache.solr.mcp.server.TestcontainersConfiguration;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
@@ -68,107 +65,4 @@ class SolrConfigTest {
 				properties.url());
 	}
 
-	@ParameterizedTest
-	@CsvSource({"http://localhost:8983, http://localhost:8983/solr",
-			"http://localhost:8983/, http://localhost:8983/solr",
-			"http://localhost:8983/solr, http://localhost:8983/solr",
-			"http://localhost:8983/solr/, http://localhost:8983/solr",
-			"http://localhost:8983/custom/solr/, http://localhost:8983/custom/solr"})
-	void testUrlNormalization(String inputUrl, String expectedUrl) {
-		// Create a test properties object
-		SolrConfigurationProperties testProperties = new SolrConfigurationProperties(inputUrl);
-
-		// Create SolrConfig instance
-		SolrConfig solrConfig = new SolrConfig();
-
-		// Test URL normalization
-		SolrClient client = solrConfig.solrClient(testProperties, new JsonResponseParser(new ObjectMapper()));
-		assertNotNull(client);
-
-		var httpClient = assertInstanceOf(Http2SolrClient.class, client);
-		assertEquals(expectedUrl, httpClient.getBaseURL());
-
-		// Clean up
-		try {
-			client.close();
-		} catch (Exception e) {
-			// Ignore close errors in test
-		}
-	}
-
-	@Test
-	void testUrlWithoutTrailingSlash() {
-		// Test URL without trailing slash branch
-		SolrConfigurationProperties testProperties = new SolrConfigurationProperties("http://localhost:8983");
-		SolrConfig solrConfig = new SolrConfig();
-
-		SolrClient client = solrConfig.solrClient(testProperties, new JsonResponseParser(new ObjectMapper()));
-		Http2SolrClient httpClient = (Http2SolrClient) client;
-
-		// Should add trailing slash and solr path
-		assertEquals("http://localhost:8983/solr", httpClient.getBaseURL());
-
-		try {
-			client.close();
-		} catch (Exception e) {
-			// Ignore close errors in test
-		}
-	}
-
-	@Test
-	void testUrlWithTrailingSlashButNoSolrPath() {
-		// Test URL with trailing slash but no solr path branch
-		SolrConfigurationProperties testProperties = new SolrConfigurationProperties("http://localhost:8983/");
-		SolrConfig solrConfig = new SolrConfig();
-
-		SolrClient client = solrConfig.solrClient(testProperties, new JsonResponseParser(new ObjectMapper()));
-		Http2SolrClient httpClient = (Http2SolrClient) client;
-
-		// Should add solr path to existing trailing slash
-		assertEquals("http://localhost:8983/solr", httpClient.getBaseURL());
-
-		try {
-			client.close();
-		} catch (Exception e) {
-			// Ignore close errors in test
-		}
-	}
-
-	@Test
-	void testUrlWithSolrPathButNoTrailingSlash() {
-		// Test URL with solr path but no trailing slash
-		SolrConfigurationProperties testProperties = new SolrConfigurationProperties("http://localhost:8983/solr");
-		SolrConfig solrConfig = new SolrConfig();
-
-		SolrClient client = solrConfig.solrClient(testProperties, new JsonResponseParser(new ObjectMapper()));
-		Http2SolrClient httpClient = (Http2SolrClient) client;
-
-		// Should add trailing slash
-		assertEquals("http://localhost:8983/solr", httpClient.getBaseURL());
-
-		try {
-			client.close();
-		} catch (Exception e) {
-			// Ignore close errors in test
-		}
-	}
-
-	@Test
-	void testUrlAlreadyProperlyFormatted() {
-		// Test URL that's already properly formatted
-		SolrConfigurationProperties testProperties = new SolrConfigurationProperties("http://localhost:8983/solr/");
-		SolrConfig solrConfig = new SolrConfig();
-
-		SolrClient client = solrConfig.solrClient(testProperties, new JsonResponseParser(new ObjectMapper()));
-		Http2SolrClient httpClient = (Http2SolrClient) client;
-
-		// Should remain unchanged
-		assertEquals("http://localhost:8983/solr", httpClient.getBaseURL());
-
-		try {
-			client.close();
-		} catch (Exception e) {
-			// Ignore close errors in test
-		}
-	}
 }
