+++ src/test/java/org/apache/solr/mcp/server/config/SolrConfigUrlNormalizationTest.java
@@ -1,0 +1,126 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.solr.mcp.server.config;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.apache.solr.client.solrj.SolrClient;
+import org.apache.solr.client.solrj.impl.Http2SolrClient;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.json.JsonTest;
+
+@JsonTest
+class SolrConfigUrlNormalizationTest {
+
+	@Autowired
+	private ObjectMapper objectMapper;
+
+	@ParameterizedTest
+	@CsvSource({"http://localhost:8983, http://localhost:8983/solr",
+			"http://localhost:8983/, http://localhost:8983/solr",
+			"http://localhost:8983/solr, http://localhost:8983/solr",
+			"http://localhost:8983/solr/, http://localhost:8983/solr",
+			"http://localhost:8983/custom/solr/, http://localhost:8983/custom/solr"})
+	void testUrlNormalization(String inputUrl, String expectedUrl) {
+		SolrConfigurationProperties testProperties = new SolrConfigurationProperties(inputUrl);
+		SolrConfig solrConfig = new SolrConfig();
+
+		SolrClient client = solrConfig.solrClient(testProperties, new JsonResponseParser(objectMapper));
+		assertNotNull(client);
+
+		var httpClient = assertInstanceOf(Http2SolrClient.class, client);
+		assertEquals(expectedUrl, httpClient.getBaseURL());
+
+		try {
+			client.close();
+		} catch (Exception e) {
+			// Ignore close errors in test
+		}
+	}
+
+	@Test
+	void testUrlWithoutTrailingSlash() {
+		SolrConfigurationProperties testProperties = new SolrConfigurationProperties("http://localhost:8983");
+		SolrConfig solrConfig = new SolrConfig();
+
+		SolrClient client = solrConfig.solrClient(testProperties, new JsonResponseParser(objectMapper));
+		Http2SolrClient httpClient = (Http2SolrClient) client;
+
+		assertEquals("http://localhost:8983/solr", httpClient.getBaseURL());
+
+		try {
+			client.close();
+		} catch (Exception e) {
+			// Ignore close errors in test
+		}
+	}
+
+	@Test
+	void testUrlWithTrailingSlashButNoSolrPath() {
+		SolrConfigurationProperties testProperties = new SolrConfigurationProperties("http://localhost:8983/");
+		SolrConfig solrConfig = new SolrConfig();
+
+		SolrClient client = solrConfig.solrClient(testProperties, new JsonResponseParser(objectMapper));
+		Http2SolrClient httpClient = (Http2SolrClient) client;
+
+		assertEquals("http://localhost:8983/solr", httpClient.getBaseURL());
+
+		try {
+			client.close();
+		} catch (Exception e) {
+			// Ignore close errors in test
+		}
+	}
+
+	@Test
+	void testUrlWithSolrPathButNoTrailingSlash() {
+		SolrConfigurationProperties testProperties = new SolrConfigurationProperties("http://localhost:8983/solr");
+		SolrConfig solrConfig = new SolrConfig();
+
+		SolrClient client = solrConfig.solrClient(testProperties, new JsonResponseParser(objectMapper));
+		Http2SolrClient httpClient = (Http2SolrClient) client;
+
+		assertEquals("http://localhost:8983/solr", httpClient.getBaseURL());
+
+		try {
+			client.close();
+		} catch (Exception e) {
+			// Ignore close errors in test
+		}
+	}
+
+	@Test
+	void testUrlAlreadyProperlyFormatted() {
+		SolrConfigurationProperties testProperties = new SolrConfigurationProperties("http://localhost:8983/solr/");
+		SolrConfig solrConfig = new SolrConfig();
+
+		SolrClient client = solrConfig.solrClient(testProperties, new JsonResponseParser(objectMapper));
+		Http2SolrClient httpClient = (Http2SolrClient) client;
+
+		assertEquals("http://localhost:8983/solr", httpClient.getBaseURL());
+
+		try {
+			client.close();
+		} catch (Exception e) {
+			// Ignore close errors in test
+		}
+	}
+}
